Update supply-chain-education-leaflet-version-2_en.md
This section contains many errors ("the copyright holder of open source grants others the right to freely use any patents that are practiced by the software and owned by the copyright holder"). 
Freely use? No, the license has conditions and requirements; the license examples don't grant a patent license to practice those patents elsewhere. Any patents practiced by the software? No, the examples are limited to the contributions in the example. etc.
I propose a simplified correction that avoids the errors.